8.96

+++ src/main/java/jp/co/toshiba/ppok/entity/Country.java
@@ -40,7 +40,7 @@ public class Country implements Serializable {
 	/**
 	 * This field corresponds to the database column INDEPENDENCE_YEAR
 	 */
-	private Integer independenceYear;
+	private Long independenceYear;
 
 	/**
 	 * This field corresponds to the database column POPULATION
@@ -50,7 +50,7 @@ public class Country implements Serializable {
 	/**
 	 * This field corresponds to the database column LIFE_EXPECTANCY
 	 */
-	private Integer lifeExpectancy;
+	private Long lifeExpectancy;
 
 	/**
 	 * This field corresponds to the database column GNP
@@ -80,7 +80,7 @@ public class Country implements Serializable {
 	/**
 	 * This field corresponds to the database column CAPITAL
 	 */
-	private Integer capital;
+	private Long capital;
 
 	/**
 	 * This field corresponds to the database column CODE2
@@ -194,7 +194,7 @@ public class Country implements Serializable {
 	 *
 	 * @return independenceYear
 	 */
-	public Integer getIndependenceYear() {
+	public Long getIndependenceYear() {
 		return this.independenceYear;
 	}
 
@@ -203,7 +203,7 @@ public class Country implements Serializable {
 	 *
 	 * @param independenceYear セットする independenceYear
 	 */
-	public void setIndependenceYear(final Integer independenceYear) {
+	public void setIndependenceYear(final Long independenceYear) {
 		this.independenceYear = independenceYear;
 	}
 
@@ -230,7 +230,7 @@ public class Country implements Serializable {
 	 *
 	 * @return lifeExpectancy
 	 */
-	public Integer getLifeExpectancy() {
+	public Long getLifeExpectancy() {
 		return this.lifeExpectancy;
 	}
 
@@ -239,7 +239,7 @@ public class Country implements Serializable {
 	 *
 	 * @param lifeExpectancy セットする lifeExpectancy
 	 */
-	public void setLifeExpectancy(final Integer lifeExpectancy) {
+	public void setLifeExpectancy(final Long lifeExpectancy) {
 		this.lifeExpectancy = lifeExpectancy;
 	}
 
@@ -338,7 +338,7 @@ public class Country implements Serializable {
 	 *
 	 * @return capital
 	 */
-	public Integer getCapital() {
+	public Long getCapital() {
 		return this.capital;
 	}
 
@@ -347,7 +347,7 @@ public class Country implements Serializable {
 	 *
 	 * @param capital セットする capital
 	 */
-	public void setCapital(final Integer capital) {
+	public void setCapital(final Long capital) {
 		this.capital = capital;
 	}
